+++ qmtl/io/datafetcher.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+"""Re-export :class:`~qmtl.sdk.data_io.DataFetcher`."""
+
+from qmtl.sdk.data_io import DataFetcher
+
+__all__ = ["DataFetcher"]
