--- qmtl/streams/__init__.py
@@ -1,5 +1,0 @@
-"""StreamInput implementations for common data sources."""
-
-from .binance import BinanceBTCStream
-
-__all__ = ["BinanceBTCStream"]
